Migrating data classes to games
