+++ mlbstatsapi/models/game/livedata/plays/play/playevent/__init__.py
@@ -1,2 +1,1 @@
 from .playevent import PlayEvent
-from .playevent import PlayCount
